requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.9
django-allauth==0.41.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
wrapt==1.11.2
